+++ addon/rules/after-initial-render.js
@@ -1,5 +1,0 @@
-export default function afterInitialRender({ firstTime, use }) {
-  if (!firstTime) {
-    return use;
-  }
-}