+++ addon/rules/always.js
@@ -1,3 +1,0 @@
-export default function always({ use }) {
-  return use;
-}