+++ tests/dummy/app/components/animated-text.js
@@ -1,9 +1,7 @@
 import Component from '@ember/component';
 import move from 'ember-animated/motions/move';
-import always from 'ember-animated/rules/always';
 
 export default Component.extend({
-  always,
   tagName: '',
   transition,
 }).reopenClass({