+++ tests/dummy/app/controllers/hero/index.js
@@ -1,6 +1,5 @@
 import Controller from '@ember/controller';
 import opacity from 'ember-animated/motions/opacity';
-import always from 'ember-animated/rules/always';
 
 function * transition({ insertedSprites, receivedSprites, removedSprites }) {
   insertedSprites.forEach(sprite => {
@@ -17,6 +16,5 @@ function * transition({ insertedSprites, receivedSprites, removedSprites }) {
 }
 
 export default Controller.extend({
-  transition,
-  always
+  transition
 });